ksmbd-tools: do full smbconf parse on ENOENT in adduser/addshare
Even if the smbconf file does not exist, we're still interested in the
`guest account' global parameter in adduser. In addshare the same would
be true for the `root directory' global parameter if it had a default
value. Do full smbconf parse even on ENOENT. Although now unnecessary,
add a check to prevent NULL deref of `global_conf.guest_account' in
__is_transient_user().

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>