ksmbd-tools: fix encode_negTokenTarg() return value
Kerberos 5 authentication requests have been prone to intermittent
failures. This is due to encode_negTokenTarg() not having a defined
return value on the success code path, which then leads to a likely
authentication error in spnego_handle_authen_request(). Fix this by
returning 0 on success. This bug was introduced when
encode_negTokenTarg() was first added in commit 9aacb96.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>