ksmbd-tools: fix caseless keys in key-value pairs
Keys are caseless according to cp_key_cmp(), but the key-value hash
table does not take that into account. This means that each case
permutation of the key is considered unique in the hash table. Fix
this by setting the GHashFunc and GEqualFunc of the GHashTable to
caseless variants. This bug has always been present.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>