ksmbd-tools: log more user information
Commit 8800eb4 established that supplementary GIDs are interesting
enough to be debug logged. The same reasoning must then apply for the
primary GID. The UID is of interest then as well especially when the
user maps to the invalid UID. Do this logging in usm_add_new_user().
Bring logging in usm_update_user_password() more in line with others.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>